refactor(popover): placeholders structure
