--- ui-admin/upload/src/placeholder-accept/index.ts
@@ -1,1 +1,0 @@
-export * from './placeholder-accept.component.js'
