--- ui-admin/upload/src/placeholder-accept/placeholder-accept.component.tsx
@@ -1,16 +1,0 @@
-import type { ReactNode }              from 'react'
-import type { HTMLAttributes }         from 'react'
-
-import { clsx }                        from 'clsx'
-
-import { basePlaceholderAcceptStyles } from './placeholder-accept.css.js'
-
-export const PlaceholderAccept = ({
-  children,
-  className,
-  ...props
-}: HTMLAttributes<HTMLSpanElement>): ReactNode => (
-  <span className={clsx(className, basePlaceholderAcceptStyles)} {...props}>
-    {children}
-  </span>
-)
