--- ui-admin/upload/src/placeholder-icon/index.ts
@@ -1,1 +1,0 @@
-export * from './placeholder-icon.component.js'
