--- ui-admin/upload/src/placeholder-link/index.ts
@@ -1,1 +1,0 @@
-export * from './placeholder-link.component.js'
