+++ ui-admin/upload/src/placeholder/accept/index.ts
@@ -1,0 +1,1 @@
+export * from './component.js'
