+++ ui-admin/upload/src/placeholder/accept/styles.css.ts
@@ -2,10 +2,8 @@ import { style } from '@vanilla-extract/css'
 
 import { vars }  from '@atls-ui-admin/theme'
 
-export const basePlaceholderLinkStyles = style({
+export const baseAcceptStyles = style({
   padding: `${vars.space.g8} ${vars.space.zero}`,
-  // @ts-expect-error correct text.blue field
-  color: vars.colors['text.blue'],
-  textDecoration: 'underline',
+  color: vars.colors['icons.placeholder'],
   fontFamily: vars.fonts.primary,
 })
