+++ ui-admin/upload/src/placeholder/dropzone/index.ts
@@ -1,0 +1,2 @@
+export type * from './interfaces.js'
+export * from './component.js'
