+++ ui-admin/upload/src/placeholder/icon/index.ts
@@ -1,0 +1,1 @@
+export * from './component.js'
