+++ ui-admin/upload/src/placeholder/icon/styles.css.ts
@@ -2,7 +2,7 @@ import { style } from '@vanilla-extract/css'
 
 import { vars }  from '@atls-ui-admin/theme'
 
-export const basePlaceholderIconStyles = style({
+export const baseIconStyles = style({
   padding: `${vars.space.g4} ${vars.space.zero}`,
   color: vars.colors['icons.placeholder'],
   fill: vars.colors['icons.placeholder'],
