+++ ui-admin/upload/src/placeholder/link/component.tsx
@@ -3,14 +3,14 @@ import type { AnchorHTMLAttributes } from 'react'
 
 import { clsx }                      from 'clsx'
 
-import { basePlaceholderLinkStyles } from './placeholder-link.css.js'
+import { baseLinkStyles }            from './styles.css.js'
 
 export const PlaceholderLink = ({
   children,
   className,
   ...props
 }: AnchorHTMLAttributes<HTMLAnchorElement>): ReactNode => (
-  <a className={clsx(className, basePlaceholderLinkStyles)} {...props}>
+  <a className={clsx(className, baseLinkStyles)} {...props}>
     {children}
   </a>
 )
