+++ ui-admin/upload/src/placeholder/link/index.ts
@@ -1,0 +1,1 @@
+export * from './component.js'
