+++ ui-admin/upload/src/placeholder/styles.css.ts
